metabase: Move `build_dbt_daily()` into a more appropriate module
